refactor: Take PR into account
